+++ go/vt/vitessdriver/convert.go
@@ -43,10 +43,8 @@ func (cv *converter) ToNative(v sqltypes.Value) (any, error) {
 		return v.ToUint64()
 	case v.IsFloat():
 		return v.ToFloat64()
-	case v.Type() == sqltypes.Datetime, v.Type() == sqltypes.Timestamp:
-		return datetimeToNative(v, cv.location)
-	case v.Type() == sqltypes.Date:
-		return dateToNative(v, cv.location)
+	case v.Type() == sqltypes.Datetime, v.Type() == sqltypes.Timestamp, v.Type() == sqltypes.Date:
+		return v.ToTimeInLocation(cv.location)
 	case v.IsQuoted() || v.Type() == sqltypes.Bit || v.Type() == sqltypes.Decimal:
 		out, err = v.ToBytes()
 	case v.Type() == sqltypes.Expression:
